src/test/java/edu/iga/adi/sm/core/direction/productions/initialization/AyTest.java
package edu.iga.adi.sm.core.direction.productions.initialization;

import edu.iga.adi.sm.TestData;
import edu.iga.adi.sm.core.Mesh;
import edu.iga.adi.sm.core.direction.Vertex;
import org.junit.Test;

import static edu.iga.adi.sm.core.Mesh.aMesh;
import static java.util.Arrays.deepToString;
import static org.assertj.core.api.Assertions.assertThat;

public class AyTest {

    private static final int GRID_SIZE = 12;

    private Mesh DUMMY_MESH = aMesh()
            .withElementsX(GRID_SIZE)
            .withElementsY(GRID_SIZE)
            .withResolutionX(12d)
            .withResolutionY(12d)
            .withOrder(2).build();

    @Test
    public void fillsWithSolution() {
        Vertex leafNode = Vertex.aVertex()
                .withBeginning(0)
                .withEnding(GRID_SIZE)
                .withMesh(DUMMY_MESH).build();

        Ay.builder()
                .node(leafNode)
                .partition(TestData.generatePartition(GRID_SIZE))
                .solution(TestData.generate2DMatrix(DUMMY_MESH.getDofsX() + 1, DUMMY_MESH.getDofsY() + 1))
                .coefficients(ExplicitMethodCoefficients.EXPLICIT_METHOD_COEFFICIENTS)
                .build()
                .apply(leafNode);

        assertThat(deepToString(leafNode.m_b)).isEqualTo(
                "[[0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0], [0.0, 8.0, 10.0, 12.0, 14.0, 16.0, 18.0, 20.0, 22.0, 24.0, 26.0, 28.0, 30.0, 32.0, 34.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]]");
    }


}